Dynamic Lighting
Sprite in front the camera darkens the scene, raycasts draw shapes within the scene and cut them out of the darkness sprite using a shader.
To use: put the FieldOfView sprite on a guard, add a mesh filter (as a child) & mesh renderer, and set the material to FoVMat, then set the mesh in the script. Set angle and radius, set target mask to 'player' and obstacle mask to 'obstacles,' set mesh resolution to 1, edge resolve iterations to ~4, and edge dist threshold to ~.5